Improve mechanism for async tests (> PY36 only)
